Updated DatSelectorAIO and added to Entropy page
